requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking -o requirements.txt
affine==2.4.0
    # via rasterio
annotated-types==0.7.0
    # via pydantic
attrs==25.3.0
    # via
    #   cfgrib
    #   eccodes
    #   ecmwf-datastores-client
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   rasterio
    #   referencing
cdsapi==0.7.6
    # via hdx-scraper-ecmwf (pyproject.toml)
certifi==2025.8.3
    # via
    #   pyogrio
    #   pyproj
    #   rasterio
    #   requests
cffi==1.17.1
    # via eccodes
cfgrib==0.9.15.0
    # via hdx-scraper-ecmwf (pyproject.toml)
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.3
    # via requests
ckanapi==4.8
    # via hdx-python-api
click==8.2.1
    # via
    #   cfgrib
    #   click-plugins
    #   cligj
    #   dask
    #   rasterio
    #   typer
click-plugins==1.1.1.2
    # via rasterio
cligj==0.7.2
    # via rasterio
cloudpickle==3.1.1
    # via dask
dask==2025.7.0
    # via hdx-scraper-ecmwf (pyproject.toml)
defopt==7.0.0
    # via hdx-python-api
dnspython==2.7.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
docutils==0.22
    # via defopt
eccodes==2.43.0
    # via cfgrib
eccodeslib==2.43.0
    # via eccodes
eckitlib==1.31.4
    # via
    #   eccodeslib
    #   fckitlib
ecmwf-datastores-client==0.4.0
    # via cdsapi
email-validator==2.2.0
    # via hdx-python-api
et-xmlfile==2.0.0
    # via openpyxl
fckitlib==0.14.0
    # via eccodeslib
findlibs==0.1.2
    # via eccodes
frictionless==5.18.1
    # via hdx-python-utilities
fsspec==2025.7.0
    # via dask
geopandas==1.1.1
    # via hdx-scraper-ecmwf (pyproject.toml)
hdx-python-api==6.4.7
    # via hdx-scraper-ecmwf (pyproject.toml)
hdx-python-country==3.9.6
    # via
    #   hdx-scraper-ecmwf (pyproject.toml)
    #   hdx-python-api
hdx-python-utilities==3.9.0
    # via
    #   hdx-scraper-ecmwf (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-country
humanize==4.12.3
    # via frictionless
idna==3.10
    # via
    #   email-validator
    #   requests
ijson==3.4.0
    # via hdx-python-utilities
inflect==7.5.0
    # via quantulum3
isodate==0.7.2
    # via frictionless
jinja2==3.1.6
    # via frictionless
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.7.0
    # via libhxl
jsonschema==4.25.1
    # via
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2025.4.1
    # via jsonschema
libhxl==5.2.2
    # via
    #   hdx-python-api
    #   hdx-python-country
locket==1.0.0
    # via partd
loguru==0.7.3
    # via hdx-python-utilities
makefun==1.16.0
    # via hdx-python-api
markdown-it-py==4.0.0
    # via rich
marko==2.2.0
    # via frictionless
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via inflect
multiurl==0.3.7
    # via ecmwf-datastores-client
num2words==0.5.14
    # via quantulum3
numpy==2.3.2
    # via
    #   cfgrib
    #   eccodes
    #   geopandas
    #   pandas
    #   pyogrio
    #   rasterio
    #   rioxarray
    #   shapely
    #   xarray
openpyxl==3.1.5
    # via hdx-python-utilities
packaging==25.0
    # via
    #   dask
    #   geopandas
    #   pyogrio
    #   rioxarray
    #   xarray
pandas==2.3.2
    # via
    #   geopandas
    #   xarray
partd==1.4.2
    # via dask
petl==1.7.17
    # via frictionless
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via sphinxcontrib-napoleon
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via frictionless
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pyogrio==0.11.1
    # via geopandas
pyparsing==3.2.3
    # via rasterio
pyphonetics==0.5.3
    # via hdx-python-utilities
pyproj==3.7.2
    # via
    #   geopandas
    #   rioxarray
python-dateutil==2.9.0.post0
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
    #   multiurl
    #   pandas
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via
    #   ckanapi
    #   frictionless
pytz==2025.2
    # via
    #   multiurl
    #   pandas
pyyaml==6.0.2
    # via
    #   dask
    #   frictionless
    #   tableschema-to-template
quantulum3==0.9.2
    # via hdx-python-api
rasterio==1.4.3
    # via rioxarray
ratelimit==2.2.1
    # via hdx-python-utilities
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   cdsapi
    #   ckanapi
    #   ecmwf-datastores-client
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   multiurl
    #   requests-file
requests-file==2.1.0
    # via hdx-python-utilities
rfc3986==2.0.0
    # via frictionless
rich==14.1.0
    # via typer
rioxarray==0.19.0
    # via hdx-scraper-ecmwf (pyproject.toml)
rpds-py==0.27.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.15
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
setuptools==80.9.0
    # via ckanapi
shapely==2.1.1
    # via geopandas
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
simplejson==3.20.1
    # via ckanapi
six==1.17.0
    # via
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
sphinxcontrib-napoleon==0.7
    # via defopt
structlog==25.4.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
tenacity==9.1.2
    # via hdx-python-country
text-unidecode==1.3
    # via python-slugify
toolz==1.0.0
    # via
    #   dask
    #   partd
tqdm==4.67.1
    # via
    #   cdsapi
    #   multiurl
typeguard==4.4.4
    # via inflect
typer==0.16.1
    # via frictionless
typing-extensions==4.14.1
    # via
    #   ecmwf-datastores-client
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   typeguard
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
unidecode==1.4.0
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.5.0
    # via
    #   libhxl
    #   requests
validators==0.35.0
    # via frictionless
wheel==0.45.1
    # via libhxl
xarray==2025.8.0
    # via
    #   hdx-scraper-ecmwf (pyproject.toml)
    #   rioxarray
xlrd==2.0.2
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsx2csv==0.8.4
    # via hdx-python-utilities
xlsxwriter==3.2.5
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities
